LPD-17786 announcements-web: Split logic into AnnouncementsAdminViewDisplayContext and AnnouncementsAdminViewManagementToolbarDisplayContext
