+++ modules/apps/announcements/announcements-web/src/main/resources/META-INF/resources/announcements_admin/init.jsp
@@ -7,4 +7,6 @@
 
 <%@ include file="/init.jsp" %>
 
+<%@ page import="com.liferay.announcements.web.internal.display.context.AnnouncementsAdminViewDisplayContext" %>
+
 <%@ include file="/announcements_admin/init-ext.jsp" %>
